+++ src/main/java/org/openlmis/requisition/web/RequisitionController.java
@@ -28,6 +28,7 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.openlmis.requisition.domain.RequisitionTemplate;
 import org.openlmis.requisition.domain.requisition.Requisition;
@@ -276,6 +277,7 @@ public class RequisitionController extends BaseRequisitionController {
   @ResponseBody
   public RequisitionDto updateRequisition(@RequestBody RequisitionDto requisitionDto,
                                           @PathVariable("id") UUID requisitionId,
+                                          HttpServletRequest request,
                                           HttpServletResponse response) {
     Profiler profiler = getProfiler("UPDATE_REQUISITION", requisitionId, requisitionDto);
 
@@ -289,6 +291,10 @@ public class RequisitionController extends BaseRequisitionController {
         () -> requisitionService.validateCanSaveRequisition(requisitionToUpdate)
     );
 
+    profiler.start("VALIDATE_VERSION");
+    requisitionVersionValidator.validateEtagVersionIfPresent(request, requisitionToUpdate)
+        .throwExceptionIfHasErrors();
+
     Map<UUID, OrderableDto> orderables = findOrderables(
         profiler, requisitionToUpdate::getAllOrderableIds
     );
@@ -298,10 +304,6 @@ public class RequisitionController extends BaseRequisitionController {
         requisitionToUpdate.getTemplate(), requisitionToUpdate.getProgramId(),
         requisitionToUpdate.getStatus(), orderables);
     requisition.setId(requisitionId);
-
-    profiler.start("VALIDATE_TIMESTAMPS");
-    requisitionVersionValidator.validateRequisitionTimestamps(requisition, requisitionToUpdate)
-        .throwExceptionIfHasErrors();
 
     ProgramDto program = findProgram(requisitionToUpdate.getProgramId(), profiler);